Hotfixes (#119)
* Merged fixes for Status of Formative Action bugs

* User story13580 (#93)

* new view for the payments to the teachers

* In the payment Teachers the amount can be inputed, also the dates cannot be in the future

* Hotfixes (#94)

* Fixed JUnit tests for the Date

* Added description to the invoices and payments

* Officially changed "Invoice" to "Movement" (#95)

* Renamed "Invoice" to "Movement"

* Fixed bug about multiple sessions setting off warnings

* Update View.java

* [UserStory13913] - Multiple Teachers (#97)

* Fixed cancellations amount payed query

* Started with user story

* Fixed Enrollment Cash

* Added comments

* Removed COIIPA's info from the view of Canceling a Formative Action

* Added custom event listeners

* Added multiple teachers for a formative action

* Bugfixes and change of the data to adapt it to new request (#98)

* after a payment is made the amount text field is set to 0 ( trying to avoid bugs). Also if the secretary clicks on 'no' in the paymnets warning the payment is made without returning  in that moment the difference

Chanes in the controller of Register Payments

* Transitioned to the new form of refunds

In the process, the nasty bug Marek discovered was fixed

* Added invoice description to the view

* Fixed model

* Added testing data

* Fixed editable

* Fixed silly thing about minimum payment

* Fixed a couple things

* Fixed problems with the dates

Payments cannot be done in the future . Also if the payment is done after the 48 h if the are places available, the places will be asigned. It checks if the cash payments are higher than 1000

* change view of registerPayment

Also 1º version of the refunds made by rosa to the professionals ( it does not work properly)

* added two textFileds to shown the due amounts


* Fixes pre-sprint review

---------

+++ src/main/java/Utils/SwingMain.java
@@ -33,7 +33,9 @@ public class SwingMain {
 			public void run() {
 				try {
 					File f= new File("database.db");
-					f.delete();
+					
+					if(f.exists())
+						f.delete();
 					
 					SwingMain window = new SwingMain();
 					window.frame.setVisible(true);
